ai_ops_assistant/requirements.txt
# AI Operations Assistant Requirements

# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3

# LLM providers (install based on your choice)
openai==1.10.0
google-generativeai==0.3.2
groq==0.4.2

# Optional: FastAPI for API mode
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
